+++ app/src/main/java/com/elias/spark/customer/domain/exception/CustomerNotFoundException.java
@@ -4,7 +4,7 @@ public class CustomerNotFoundException extends RuntimeException {
 
 	private static final long serialVersionUID = 1L;
 
-	public CustomerNotFoundException(Long id) {
+	public CustomerNotFoundException(Integer id) {
 		super("Customer with id " + id + " not found.");
 	}
 }
